TimeParallel.jl

This is the documentation of TimeParallel.jl, a Julia package implementing time-parallel methods.

Manual

TimeParallel.jl

Manual

API

Types

Functions

Solvers

Utilities

Index

API

All exported types and functions are considered part of the public API and thus documented in this manual.

Types

Missing docstring.

Missing docstring for ErrorCheck. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TimeParallelSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TimeParallelIterate. Check Documenter's build log for details.

TimeParallel.TimeParallelSolution — Type

TimeParallelSolution{iterates_T, φ_T, U_T, T_T} <: InitialValueSolution

returns a constructor for the numerical solution of an InitialValueProblem obtained with a TimeParallelSolver.

TimeParallelSolution(problem, solver::TimeParallelSolver)

returns an initialised TimeParallelSolution given a problem, e.g. an InitialValueProblem, and a TimeParallelSolver.

Functions

NSDEBase.solve — Function

solve(problem, solver::TimeParallelSolver; mode::String = "SERIAL")

returns the TimeParallelSolution of a problem, e.g. an InitialValueProblem.

Missing docstring.

Missing docstring for solve!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coarseguess!. Check Documenter's build log for details.

Solvers

TimeParallel.Parareal — Type

Parareal{error_check_T, ℱ_T, 𝒢_T, P_T, K_T} <: TimeParallelSolver

returns a constructor for the TimeParallelSolver based on the parareal algorithm.

Parareal(ℱ::Function, 𝒢::Function; P = 10, K = P, 𝜑 = 𝜑₁, ϵ = 1e-12)

returns a Parareal with:

ℱ :: Function : fine solver.

𝒢 :: Function : coarse solver.

P :: Integer : number of time chunks.

K :: Integer : maximum number of iterations.

𝜑 :: Function : error control function.

ϵ :: Real : tolerance.

Parareal(finesolver::InitialValueSolver, coarsolver::InitialValueSolver; P = 10, K = P, 𝜑 = 𝜑₁, ϵ = 1e-12)

returns a Parareal from a finesolver and a coarsolver.

Utilities

Missing docstring.

Missing docstring for getchunks. Check Documenter's build log for details.

Index

TimeParallel.Parareal

TimeParallel.TimeParallelSolution

NSDEBase.solve