TimeParallel.jl

This is the documentation of TimeParallel.jl, a Julia package implementing time-parallel methods.

Manual

TimeParallel.jl

Manual

API

Types

Functions

Solvers

Utilities

Index

API

All exported types and functions are considered part of the public API and thus documented in this manual.

Types

Missing docstring.

Missing docstring for ErrorCheck. Check Documenter's build log for details.

TimeParallel.TimeParallelSolver — Type

An abstract type for time-parallel solvers.

Missing docstring.

Missing docstring for TimeParallelIterate. Check Documenter's build log for details.

TimeParallel.TimeParallelSolution — Type

TimeParallelSolution <: InitialValueSolution

A composite type for the solution of an InitialValueProblem obtained with a TimeParallelSolver.

Constructors

TimeParallelSolution(problem, solver)

Arguments

problem : initial value problem, e.g. an InitialValueProblem.

solver :: TimeParallelSolver.

Functions

getindex : get iterate.

lastindex : last index.

length : number of iterates.

setindex! : set iterate.

show : shows name and contents.

summary : shows name.

Functions

NSDEBase.solve — Function

solve(problem, solver::TimeParallelSolver; kwargs...) :: TimeParallelSolution

returns the TimeParallelSolution of a problem, e.g. an InitialValueProblem.

NSDEBase.solve! — Function

solve!(solution::TimeParallelSolution, problem, solver::TimeParallelSolver; kwargs...) :: TimeParallelSolution

returns the TimeParallelSolution of a problem, e.g. an InitialValueProblem.

Missing docstring.

Missing docstring for coarseguess!. Check Documenter's build log for details.

Solvers

TimeParallel.Parareal — Type

Parareal <: TimeParallelSolver

A composite type for the parareal algorithm.

Constructors

Parareal(ℱ, 𝒢; P = 10, K = P, φ = ErrorCheck(; 𝜑 = 𝜑₁, ϵ = 1e-12))
Parareal(ℱ, 𝒢; P = 10, K = P, 𝜑 = 𝜑₁, ϵ = 1e-12)

Arguments

ℱ :: Union{Function, InitialValueSolver} : fine solver.

𝒢 :: Union{Function, InitialValueSolver} : coarse solver.

P :: Integer : number of time chunks.

K :: Integer : maximum number of iterations.

𝜑 :: Function : error control function.

ϵ :: Real : tolerance.

Functions

show : shows name and contents.

summary : shows name.

Utilities

Missing docstring.

Missing docstring for getchunks. Check Documenter's build log for details.

Index

TimeParallel.Parareal

TimeParallel.TimeParallelSolution

TimeParallel.TimeParallelSolver

NSDEBase.solve

NSDEBase.solve!